IT24103717Lab4Q1.java

+++ IT24103717Lab4Q1.java
@@ -6,7 +6,7 @@ public class IT24103717Lab4Q1{
  Scanner input= new Scanner(System.in);
 
 System.out.print("Enter a number: ");
-double number = input.nextDouble();
+int number = input.nextInt();
 
 if (number > 0) {
     System.out.print("The number is: Positive");
